+++ plone/app/tiles/browser/base.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from AccessControl import Unauthorized
 from plone.autoform.form import AutoExtensibleForm
+from plone.z3cform.fieldsets.group import Group
 from plone.z3cform.interfaces import IDeferSecurityCheck
 from z3c.form.interfaces import IWidgets
 from zope.component import getMultiAdapter
@@ -27,6 +28,29 @@ class TileFormLayout(object):
         return self.index.macros
 
 
+class TileFormGroup(Group):
+    """TileForm group class, which prefixes group form fields similarly
+    to TileForm"""
+
+    prefix = ''
+
+    def updateWidgets(self, prefix=None):
+        # Override to set the widgets prefix before widgets are updated
+        # Note: updateWidgets(prefix=...) is not yet supported by
+        # z3c.form 2.5.1 used by Plone 4.2
+        self.widgets = getMultiAdapter(
+            (self, self.request, self.getContent()), IWidgets)
+        for attrName in ('mode', 'ignoreRequest', 'ignoreContext',
+                         'ignoreReadonly'):
+            value = getattr(self.parentForm.widgets, attrName)
+            setattr(self.widgets, attrName, value)
+        if prefix is not None:
+            self.widgets.prefix = prefix
+        else:
+            self.widgets.prefix = self.parentForm.tileType.__name__
+        self.widgets.update()
+
+
 class TileForm(AutoExtensibleForm):
     """Mixin class for tile add/edit forms, which will load the tile schema
     and set up an appropriate form.
@@ -38,11 +62,14 @@ class TileForm(AutoExtensibleForm):
 
     # Get fieldets if subclass sets additional_schemata
     autoGroups = True
+    group_class = TileFormGroup
 
-    # We don't want the tile edit screens to use a form prefix, so that we
-    # can pass simple things on the edit screen and have them be interpreted
-    # by transient tiles
+    # We set prefix dynamically in updateWidgets
     prefix = ''
+    # Note: We used to not want the tile edit screens to use a form prefix, so
+    # that we could pass simple things on the edit screen and have them be
+    # interpreted by transient tiles, but this was deprecated without no
+    # obvious reason in 85bff714b161eca07b66736780f293940f4f1d92
 
     # Name is used to form the form action url
     name = ''
@@ -77,11 +104,13 @@ class TileForm(AutoExtensibleForm):
             return ''
         return super(TileForm, self).render()
 
-    def updateWidgets(self):
+    def updateWidgets(self, prefix=None):
         # Override to set the widgets prefix before widgets are updated
+        # Note: updateWidgets(prefix=...) is not yet supported by
+        # z3c.form 2.5.1 used by Plone 4.2
         self.widgets = getMultiAdapter(
             (self, self.request, self.getContent()), IWidgets)
-        self.widgets.prefix = self.tileType.__name__
+        self.widgets.prefix = prefix or self.tileType.__name__
         self.widgets.mode = self.mode
         self.widgets.ignoreContext = self.ignoreContext
         self.widgets.ignoreRequest = self.ignoreRequest